feat: move fetchParticipants to utils
